Change to work with .deb and .rpm

+++ vscode.sh
@@ -1,30 +1,30 @@
 #!/bin/bash
 
+OS="$(lsb_release -is)"
+
 clear
+
 echo "========================================================="
-echo "Baixando última versão estável amd64 em /var/tmp"
+echo " VSCODE "
 echo "========================================================="
+
 cd /var/tmp
-wget --show-progress -O code.deb https://vscode-update.azurewebsites.net/latest/linux-deb-x64/stable
 
-echo "========================================================="
-echo "Desinstalando versões anteriores"
-echo "========================================================="
-sudo apt-get -y purge code
-sudo apt-get autoremove
-sudo apt-get autoclean
+if [ "$OS" == "Ubuntu" ]; then
+    wget --show-progress -O code.deb https://vscode-update.azurewebsites.net/latest/linux-deb-x64/stable
 
-echo "========================================================="
-echo "Instalando nova versão"
-echo "========================================================="
-sudo dpkg -i code.deb
+    sudo apt-get -y purge code
+    sudo apt-get autoremove
+    sudo apt-get autoclean
+    sudo dpkg -i code.deb
+    sudo apt-get -f -y install
+    sudo rm code.deb
+fi
 
-echo "========================================================="
-echo "Instalando dependências"
-echo "========================================================="
-sudo apt-get -f -y install
+if [ "$OS" == "Fedora" ]; then
+    sudo rpm --import https://packages.microsoft.com/keys/microsoft.asc
+    sudo sh -c 'echo -e "[code]\nname=Visual Studio Code\nbaseurl=https://packages.microsoft.com/yumrepos/vscode\nenabled=1\ngpgcheck=1\ngpgkey=https://packages.microsoft.com/keys/microsoft.asc" > /etc/yum.repos.d/vscode.repo'
 
-echo "========================================================="
-echo "Removendo instalador"
-echo "========================================================="
-sudo rm code.deb
+    dnf check-update
+    sudo dnf install code
+fi
